+++ src/nbfc.h
@@ -5,6 +5,7 @@
 #define NBFC_MAX_FILE_SIZE               32768
 #define NBFC_TEMPERATURE_FILTER_TIMESPAN 6000 /*ms*/
 #define NBFC_MODEL_CONFIGS_DIR           DATADIR "/nbfc/configs"
+#define NBFC_MODEL_SUPPORT_FILE          DATADIR "/nbfc/model_support.json"
 #define NBFC_CONFIG_DIR                  SYSCONFDIR "/nbfc"
 #define NBFC_SERVICE_CONFIG              SYSCONFDIR "/nbfc/nbfc.json"
 #define NBFC_PID_FILE                    RUNSTATEDIR "/nbfc_service.pid"